fix: transfer generate jwt token to separate package
