+++ internal/entity/user_test.go
@@ -3,13 +3,7 @@ package entity
 import (
 	"crypto/sha256"
 	"encoding/hex"
-	"fmt"
 	"testing"
-	"time"
-
-	"github.com/golang-jwt/jwt/v4"
-	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 )
 
 func TestUser_CheckPassword(t *testing.T) {
@@ -146,73 +140,5 @@ func TestUser_GetPasswordHash(t *testing.T) {
 }
 
 func TestUser_GenerateJWTToken(t *testing.T) {
-	type fields struct {
-		Login    string
-		Password string
-		ID       uint
-	}
-	type args struct {
-		secretKey string
-		expTime   time.Duration
-	}
-	tests := []struct {
-		name    string
-		fields  fields
-		args    args
-		wantErr bool
-	}{
-		{
-			name: "success case",
-			fields: fields{
-				Login:    "123",
-				Password: "password",
-				ID:       10,
-			},
-			args: args{
-				secretKey: "secret-key",
-				expTime:   time.Hour * 1,
-			},
-			wantErr: false,
-		},
-		{
-			name: "negative empty secret case",
-			fields: fields{
-				Login:    "123",
-				Password: "password",
-				ID:       10,
-			},
-			args: args{
-				secretKey: "",
-				expTime:   time.Hour,
-			},
-			wantErr: true,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			u := &User{
-				Login:    tt.fields.Login,
-				Password: tt.fields.Password,
-				ID:       tt.fields.ID,
-			}
-			got, err := u.GenerateJWTToken(tt.args.secretKey, tt.args.expTime)
-			if tt.wantErr {
-				require.Error(t, err)
-				return
-			} else {
-				require.NoError(t, err)
-			}
 
-			claims := &Claims{}
-			token, err := jwt.ParseWithClaims(got, claims, func(token *jwt.Token) (interface{}, error) {
-				if _, ok := token.Method.(*jwt.SigningMethodHMAC); !ok {
-					return nil, fmt.Errorf("invalid signing method: %w", err)
-				}
-				return []byte(tt.args.secretKey), nil
-			})
-
-			assert.True(t, token.Valid)
-			assert.Equal(t, tt.fields.ID, claims.UserID)
-		})
-	}
 }
